modules/modeling/Hadoop/sqoop1/sqoop1_import_hdfs/requirements.txt
pyDataCanvas >= 0.4.3,<0.5.0
boto
